Use partition set tag to identify partitino set runs (rather than the pipeline_name column)
Summary:
We have some evidence that suggests combining a column filter on the runs table (pipeline_name) and a filter on the tags table is very expensive due to join performance. Instead, use another tag to filter runs down to the right partition set.

Once we have a column for repo label, we could switch back to just a query on the runs table - having the condition on both tables seems to be what is causing problems.

Test Plan: Partitions set query that used to take ~20 seconds now takes <1 second