[dagit] Allow deleting a run from the Run page itself (#9729)
### Summary & Motivation

Add a menu item to the Run detail header menu, allowing deletion of the run. This uses the same deletion dialog as elsewhere, so if the run is still in progress, it will give the user the option to terminate instead.

When a run is deleted, redirect to the runs list for the job.

<img width="241" alt="Screen Shot 2022-09-19 at 1 22 38 PM" src="https://user-images.githubusercontent.com/2823852/191087525-3e377660-a127-4c36-b1cd-c3a3e767c631.png">

### How I Tested These Changes

View a completed run, verify that I can open the deletion dialog and perform the deletion.

Repeat with an in-progress run, verify that I can terminate it instead.