[dagit] Make asset group row clickable (#9810)
### Summary & Motivation

Make the asset group row in the virtualized asset table clickable. It is also focusable and openable using the keyboard, though the focus state isn't obvious and we should improve it.

### How I Tested These Changes

Click on the asset group row on the virtualized table. Verify that it toggles open/close. Tab to it, verify that space/enter also does this.