Make alembic pin much less restrictive
Summary:
Took a look at the test that was failing in 1.7.0 and it looks like the underlying issue was fixed in the very next version, and the latest version passes tests too. Based on that, makes the pin less restrictive.

Resolves https://github.com/dagster-io/dagster/issues/7891.

Test Plan:
BK, verify that a migration from 0.15.x to 1.0.x still works with the latest version