Unguard setup() calls
Previously, all of our `setup()` calls were guarded behind:

```
if __name__ == "__main__":
```

This makes it difficult to interact with the setup.py files via
`distutils.core.run_setup`.

As far as I can tell, it's best practice to _not_ guard these calls
because it's generally assumed that these are run as scripts and not
directly imported.

It's not used in the official docs:

https://docs.python.org/3/distutils/setupscript.html

And this random answer on StackOverflow says it's not common:

https://stackoverflow.com/questions/59831397/run-setup-function-from-setuptools-only-if-name-main

So I'm removing it for now and will revert if this causes problems. If
we choose to keep the guard, we'll have to get cute with how we read
requirements (monkeypatch `__name__` or jumping through some clever
`importlib`/`distutils` hoops).