Find all of the changed files
I think we'll always want to compare this against origin/master (instead
of HEAD~). This is because we want to see a completely green test suite
before merging. So even if you make a 1 line change to fix a specific
subset of tests, we'll want to rerun all of the tests affected by the
branch.

Alternatively, we could try to compare against the "parent" branch (as
in the branch that our branch was created from). But that's not really a
thing in git. It would afford us the opportunity to run smaller tests on
large stacks of diffs. But I'm punting on attempting to implement it.