[dagit] Repair timeline empty state height (#9870)
### Summary & Motivation

Because elements of the run timeline are absolutely positionined, if we don't take into account the height of all of its elements, we can end up with it overlapping subsequent contents. This currently happens in the empty state of the timeline.

Fix this by applying height appropriately.

### How I Tested These Changes

In and out of new workspace feature flag, load timeline with runs and with empty state. Verify that the contents of the timeline don't overlap other pieces of content.