[dagit] Move Code location and Daemon status polling to context (#9883)
### Summary & Motivation

We want to be able to show icons on the Deployment tabs, specifically "Code locations" and "Daemons" in open source.

In order to do this, move the status polling behavior to a context value that can be consumed by both the nav icons and the tabs on the Deployment page.

In open source, we'll poll for both `code-locations` and `daemons`. In Cloud, we only want `code-locations` because we never show icons for "Agents".

With flag, plus code location and daemon errors:

<img width="469" alt="Screen Shot 2022-10-04 at 10 30 02 AM" src="https://user-images.githubusercontent.com/2823852/193904328-f076cae4-8fe0-47de-92e3-0d7ec0696afe.png">

Without flag, with daemon errors:

<img width="385" alt="Screen Shot 2022-10-04 at 10 30 17 AM" src="https://user-images.githubusercontent.com/2823852/193904384-38b002ed-0d54-4c0e-9a1a-e7cae7f7d39a.png">


### How I Tested These Changes

In open source Dagit, verify that the warning icon and spinner appear correctly in the top nav and on the Deployment page tabs. Verify that this works correctly both with and without the "New workspace page" feature flag.

Repeat in Cloud, verify that Cloud master works correctly with this change in OS.