Only add buildkite steps for affected changes
This will benefit from some refactoring once I play around with it a bit
more.

We're still going to see a lot of tests run on a lot of PRs because
almost everything requires dagster.

Right now, this doesn't build the steps at all. We might also want to
experiment with somehow marking the steps as skipped so it's obvious
what did and didn't run on a PR.