Parametrize Buildkite queues
So we can globally switch queues without needing all in flight
branches to be rebased (obviously this benefit will only take
affect _after_ existing branches rebase to include _this_ change).

This will make it easier to migrate a bunch of builds to a new queue
without needing to keep the old queue active for a long tail of older
builds.

NO_SKIP