Return each Python PackageSpec's distribution (#9889)
* Return each Python PackageSpec's distribution

To limit each build's scope, we want to begin only running tests
affected by the change.

We can assume a package's tests need to run if:

- The package's tests change
- The package changes
- Any of the package's dependencies change (package code only - not
  their tests)

This change modifies PackageSpec to parse each package's setup.py file.
Using a combination of a git diff and the distribution's
`install_requires` and `extras_require` attributes, we should be able to
pinpoint the exact packages that are affected by each change.

`distutils.core.run_setup` does the heavy lifting here:

https://docs.python.org/3/distutils/apiref.html#distutils.core.run_setup

> Run a setup script in a somewhat controlled environment, and return
> the distutils.dist.Distribution instance that drives things. This is
> useful if you need to find out the distribution meta-data (passed as
> keyword args from script to setup()), or the contents of the config
> files or command-line.

* Unguard setup() calls

Previously, all of our `setup()` calls were guarded behind:

```
if __name__ == "__main__":
```

This makes it difficult to interact with the setup.py files via
`distutils.core.run_setup`.

As far as I can tell, it's best practice to _not_ guard these calls
because it's generally assumed that these are run as scripts and not
directly imported.

It's not used in the official docs:

https://docs.python.org/3/distutils/setupscript.html

And this random answer on StackOverflow says it's not common:

https://stackoverflow.com/questions/59831397/run-setup-function-from-setuptools-only-if-name-main

So I'm removing it for now and will revert if this causes problems. If
we choose to keep the guard, we'll have to get cute with how we read
requirements (monkeypatch `__name__` or jumping through some clever
`importlib`/`distutils` hoops).

* Use absolute paths for reading the version

Now that we're parsing these files with `distutils.core.run_setup`, we
can't rely only on relative paths for reading the version file.

* Temporarily ignore distutils deprecation

We should find a more modern approach to parsing our setup.py but I'm
just silencing the deprecation warning for now.

importlib_metadata looks promising but it appears to work only for
installed packages.