Fetch main
Because we use Buildkite's git mirror experiment, origin/master isn't
always up to date. This attempts to fetch it to make our diff more
accurate. If we actually go this route, we'll probably want to instread
provide it as an agent hook or look at using a Buildkite plugin like:

https://github.com/chronotc/monorepo-diff-buildkite-plugin