Don't raise an exception when changing the EcsRunLauncher's container name to a new name
Summary:
This fixes an issue when you change the EcsRunLauncher from expecting a container with one name to expecting a container with a different name - the reuse task definition function raises an Exception when detecting if the old task definition (with the old container name) should be re-used. Instead, it should just choose to not use the old container definition.

Test Plan: Adding a BK test now