Skip dagit builds unless necessary
This brings the dagit skipping logic inline with our new skipping logic.
Traditionally, we've always run dagit and we've attempted to skip most
of the rest of the build if we detect dagit-only changes. Now, we'll
follow the same principle as the other steps where every step is always
added to the pipeline, but each is conditionally marked as skipped
depending on whether it needs to run or not.

The logic for running the dagit JS tests is:

1. On every merge to main or the release branch, execute JS tests
2. On every change to any file in the js_modules directory, execute JS
   tests

Otherwise, mark the JS tests as skipped. The git diffing loging
introduced in #9897 more reliably identifies changed files. One small
departure from the original dagit skipping logic is that we always diff
against `main` instead of diffing against the base branch for reasons
described in:

https://github.com/dagster-io/dagster/pull/9897/commits/55333bd413f078657acca03976534febff0009dc

Although we could potentially bring that diffing behavior back in a
future PR.