Infer skips without setup.py
Previously, we relied on the existence of a setup.py to infer skips.

Now, we can also provide a requirements.txt. This is for things like our
integration test suites that never actually build a Python distribution.
Today, their dependencies are only defined in a tox.ini. We'll
instead/in addition explicitly define dependencies in a requirements.txt
as well.

This also introduces a simplifying change to the skipping logic.
Previously, we drew a distinction between packages that had their
implementation changed and packages that had their test changed. This
has been generalized so that now a package runs its steps if:

1. Any file that can drive behavior within its directory changes. This
   covers the existing case of "my tests changed" and "my implementation
   changed"
2. Any implementation file changes in an upstream package