Refactor build python package skipping (#10037)
Originally, this logic was implemented inside of a function that is
responsible for building most - but not all - of our PackageSpec steps.
This was convenient because we already had a list of most - but not all
- packages that we could iterate over.

But things like our helm steps and our integration steps still leverage
PackageSpec but were not caught with the previous change. This refactor
moves the skipping logic into PackageSpec itself.

We do this by using a glob statement to list all available packages
(regardless of whether they have PackageSpecs defined or not). We use
the existince of a setup.py file as evidence for a package - if we ever
move to a different build process, we'll need to change this logic. One
interesting check we could add to this is something to make sure every
package actually has a PackageSpec - but that's work for another day.

The same general logic applies. Python package steps run on:

1. Every merge to main
2. If the package's implementation changes
3. If the package's tests change
4. If any of the package's requirements' implementations change

It's possible we don't have our full dependency tree spelled out -
particularly for things like our integration tests - so I'll likely have
a series of follow up PRs to patch holes.