Raise if we can't infer distribution
This fixes a kind of strange issue that bit us in #10063.

If `run_setup()` doesn't actually find a `setup()` function to execute,
it's supposed to raise:

https://github.com/pypa/distutils/blob/45295fc5dc2b85a4a0d19b92429d4779db87ba79/distutils/core.py#L277-L283

However, each time it runs, it stores state in a global module variable:

https://github.com/pypa/distutils/blob/45295fc5dc2b85a4a0d19b92429d4779db87ba79/distutils/core.py#L254

This means if we call `run_setup()` on a script without a `setup()`
function (or with a `setup()` function that's guarded by
`if __name__ == "__main__":`) and we've already successfully run
`run_setup()`, we get the _previous_ distribution instead of raising an
error.

Now, we reload the module each time to clear out the global.