Remove references to services in docker-compose down command test fixture (#10042)
### Summary & Motivation
A recent change to docker-compose down errors on additional arguments that were previously ignored (https://github.com/docker/compose/pull/9158). We had our docker-compose down setup to exactly mirror docker-compose up, which is incorrect in the exact way that it seems like this PR seeks to avoid confusion, so this PR gets rid of that parity in the command.

### How I Tested These Changes
Previously got errors on local spindown when using docker-compose, this fixes.