Use instance.python_log_level when determining default system logger level (#10073)
Summary:
This makes it so that if you set python_log_level to INFO in your dagster.yaml, it applies to the console output from dagit. I originally held off on this because setting this could affect the celery_executor and docker_celery_executor and celery_k8s_executor, which still read in system logs during the execution - but given that those are off the beaten path and python_log_level is an experimental feature, I think it's OK to set this? We could land it as part of 1.1 if we think there are people out there using this log level flag as well as one of those executors.