Fix assets with dashes in their path and `io_manager_def` (#10087)
### Summary & Motivation

Dagster accepts dashes (`-`) in asset key path components. When automatically creating an operation for assets, dashes are replaced by underscores (`_`).

However, when defining an asset with either `@asset(..., io_manager_def=...)` or `SourceAsset(..., io_manager_def=...)`, it fails if the asset key contains a dash:

```
dagster._check.CheckError: Failure condition: Resource key 'my-repo_test1__io_manager' must be a valid Python identifier.
```

The reason is that the sanitization done for creating an operation is not done when creating an I/O manager resource. This PR adds a method to `AssetKey` to produce a sanitized name that can be used for the operation name as well as in `@asset` and `SourceAsset` for `io_manager_def`. This fixes the issue, and avoids hardcoding this rule in three places.

### How I Tested These Changes

This does not produce an error anymore:

```py
test_asset = SourceAsset(
   ["some-prefix", "test-asset"],
   io_manager_def=custom_io_manager
)
```