Extend default timeout
A lot of our builds run long because test suites hit their 20 minute
timeout and re-run from the beginning.

This timeout is for the buildkite step - not the actual tests. So all of
the Docker pulling, tox setup, etc. counts againt our total.

I'd like to extend this global timeout while I investigate:

  - shortening the amount of pre-work that needs to happen in each step
  - speeding up/breaking apart our longest running test suites

Anecodtally, many timeouts happen when we're > 90% done with the tests
so I think a few more minutes mightly drastically reduce the number of
Buildkite retries (and consequently, the length of most builds).

NO_SKIP=true