Parse HEAD commit instead of BUILDKITE_MESSAGE
It looks like when you trigger a build by opening a PR,
BUILDKITE_MESSAGE is the PR title and not the body of the latest
commit.

Subsequent commits set BUILDKITE_MESSAGE to the commit message.

This means if you want to use any of our substring tooling to configure
buildkite (NO_SKIP=true, etc.), you need to either include it in your
title of your PR or you need to add a second commit.

Now, we'll parse both BUILDKITE_MESSAGE and check the HEAD commit
message.