[dagit] Make 50-level colors opaque (#10186)
### Summary & Motivation

The 50-level colors in `@dagster-io/ui` are all low-opactiy values, which means that we can end up with weird states where rendering these colors in layered situations. For instance, when hovering over a `Tag` with its `data-tooltip` visible, the translucency breaks the UI:

<img width="272" alt="Screen Shot 2022-10-25 at 10 01 01 AM" src="https://user-images.githubusercontent.com/2823852/197825791-bc113a84-8c05-4b53-ba48-e267f8450183.png">

I'm not really sure that we need these to be low-opacity colors. Our primary background colors are white and light gray, and these light colors should look roughly the same on these backgrounds irrespective of whether they're rendered as opaque or translucent. This could end up being different on a darker background...but I'm not sure we would want these 50-level colors to show up as translucent on dark backgrounds anyway.

This isn't fixing anything that hasn't been broken for a while, so I think this can wait until we can discuss the design ramificiations.

With this change:

<img width="166" alt="Screen Shot 2022-10-25 at 11 08 23 AM" src="https://user-images.githubusercontent.com/2823852/197825900-9b519162-55c5-449d-8c37-b75a2063c8a4.png">


### How I Tested These Changes

View tags throughout Dagit, verify that they look fine on white and gray wash backgrounds.