Don't repeat skip logs (#10307)
This changes from using a NamedTuple to using a dataclass so we can more
easily have a few "private" instance variables to control memoization of
our skip reason property. Decorating it with an lrucache wasn't worth
the effort because our clash isn't hashable.

Instead, use two variables to decide whether or not we already have a
skip reason:

self._should_skip which initializes as empty but gets set to False any
time we have no skip reason. This is because Buildkite's API expects
either a string or nothing so self._skip_reason being None is a valid
condition.

self._skip_reason which initializes as empty but is assigned a string if
we actually do skip.

The end result is cleaner logs because we don't repeat lines every time
we check the skip reason.