Remove special casing for graphql tests (#10327)
We had a file in here called setup.py that doesn't actually behave like
a typical Python setup.py - it's just defining a repo to be used in
other tests.

We had special casing in our PythonPackages to ignore anything with
`_tests` in the filepath but that's overly restrictive. Instead, let's
rename the poorly named module.