Back-compat fix for setting the task definition in the EcsRunLauncher to an env var
Summary:
We recently moved task_definition from a StringSource to a scalar union, but some users were relying on setting the task definition to an environment variable. In hindsight we probably could have made it an entirely separate field instead of a scalar union, but for now to keep the current behavior without breaking old behavior, set an 'env' key to the dictionaryu that allows it to still be used to source the string from a environment variable.

Test Plan: BK