cached_method decorator
Docstring:

Caches the results of a method call.

Usage:

    .. code-block:: python

        class MyClass:
            @cached_method
            def fetch_value_from_database(self, key):
                ...

The main difference between this decorator and `functools.lru_cache(max_size=None)` is that each
instance of the class whose method is decorated gets its own cache. This means that the cache
can be garbage-collected when the object is garbage-collected.

A more subtle difference between this decorator and `functools.lru_cache` is that this one
prioritizes preventing mistakes over ease-of-use and performance. With `functools.lru_cache`,
these three invocations would all result in separate cache entries:

    .. code-block:: python

        class MyClass:
            @cached_method
            def a_method(self, arg1, arg2):
                ...

        obj = MyClass()
        obj.a_method(arg1="a", arg2=5)
        obj.a_method(arg2=5, arg1="a")
        obj.a_method("a", 5)

With this decorator, the first two would point to the same cache entry, and non-kwarg arguments
are not allowed.