Hook for loading secrets in grpc/run/step entry points (#10089)
RFC: hook for loading secrets in grpc/run/step entry points

Summary:
Adds a new instance component that lets user code load secrets and inject them into the environment (instead of our usual MO which is to rely on the surrounding compute environment to supply secrets, e.g. through kubernetes secrets specified in the Helm chart or injected in the run launcher).

The sketchiest part here is the addition of serialized instance ref as an argument to the grpc command - which would mean that this would only work in situations where some external Python process is managing the construction of grpc servers.