RFC: Automatically load env vars from a local .env file when one is present during dagster local development
Summary:
This makes the default dagit startup path check for a local .env file and pull secrets into it. The idea would be to make there be a dirt-simple way to get env vars into your local Dagster when getting started and doing local development.

It wouldn't work easily out of the box in docker/k8s/etc.

It would work well when doing local development with other services that export a .env file for you (like potentially Dagster Cloud!)

Will add tests and such if we think this implementation seems reasonable.

Also moved the default log level of local grpc servers to INFO so that there's some observability when it injects secrets from the file.