Pass an instance ref and location into local grpc servers
Summary:
This is to lay groundwork for https://github.com/dagster-io/dagster/pull/10491/files - fixes up a whole bunch of tests that were assuming that you could run a grpc server using an in memory instance (including a couple of whole graphql test suites), and pass in the local instance ref into the grpc server constructor. This isn't something that a user has actually been able to do for a while.

Test Plan: BK