Add build_snowflake_pandas_io_manager wrapper to eliminate use of type handlers in the common case (#10615)
### Summary & Motivation

To use our built in i/o managers for duckdb and snowflake, one has to import two libraries and be introduced to the concept of "type handlers", which are pretty arcane implementation detail that most people don't have to know about it. This PR introduces a decorated snowflake_pandas_io_manager that obviates the need to introduce SnowflakePandasTypeHandler in the common case. It also converts one example to this new function as a demo. If we move forward, we would also do this for duckdb and convert all examples.

### How I Tested These Changes

Unit tests