[definitions-stack-1] Add Definitions class entry point
Summary & Motivation:

Repository and workspace are duplicate and confusing concepts in Dagster. This is the beginning of an effort to eliminate them and consolidate under the concept of a code location.

This introduces a new `Definitions` class that can be used an as entry point in dagster's tool. Dagit, the CLI, and any other tools discover in instance of this at the scope of the module they are loading. The variable must be named `defs`. There can only be one of these loaded, so the notion of having multiple repositories in a single code location goes away.

Importantly this also makes resources a top-level concept rather than something that must be threaded to all individual definitions. This diff does that for assets currently. Later we will apply that to jobs. In the short-term it eliminates the need for with_resources calls to assets.

This also makes the top-level API typed and discoverable, rather than the current `@repository` API which returns an array of heterogenous objects that are poorly documented.

TODO, link to public github discussion on this.

Test Plan: Included unit tests

Reviewers: yuhan, sryza, OwenKephart

Reviewed By: OwenKephart

Pull Request: https://github.com/dagster-io/dagster/pull/10655