[dagit] Remove tag fetch for RunsFilterInput typeahead (#10757)
### Summary & Motivation

The query that retrieves the full list of key/value pairs that may be used for tag searching on Runs is potentially super expensive. Since the user can already use freeform typing to add a tag filter to the search, and because tags can be added to search via hover+click on tags shown in the Runs table, let's remove the tag search query altogether.

### How I Tested These Changes

View Runs, type "tag:", verify that I can complete it with "tag:foo=bar" and successfully find matching runs. Use hover+click on rendered tags in the table, verify that search works with these as well.