[dagit] Don't include repo tag when filtering snapshot runs (#10786)
### Summary & Motivation

When viewing the Runs tab in a job snapshot view, the query fails because we try to apply a filter tag that has the repo and location -- a snapshot doesn't have a code location, and the resulting tag is `undefined@undefined`.

I think we can just omit this tag in the snapshot case, because we have a snapshot ID already, and that works well enough for our filtering.

In the Runs tab for a job that is loaded, it will continue to use the repo tag to ensure that we get the correct list of runs.

### How I Tested These Changes

View Runs tab for a snapshot, verify that the query does not include a repo tag.

View Runs tab for a loaded job, verify that it does.