[code-location-concept-page-1] Concept page for code locations [WIP]
Summary & Motivation: This is a WIP, but I wanted to throw up the intro verbiage
to kickstart the conversation here.

The decision I want to discuss here is separating the notion of a python module the has a `Definitions` object defined at module-scope in the `defs` from an entity that is loadable by our tools. This is a subtle but important point. It means that a location doesn't really exist until it it loadable by our tools. So when you type `dagit -f foo.py` it is meant to mean "I am loading this code location in dagit with the current Python environment, communicating with that Python Environment locally, and loading the Python file foo.py as the entry point". This is different than saying that foo.py is a code location, even though undoubtedly it will be referred to it as that colloquially.

Test Plan: Read

Reviewers: erinkcochran87

Pull Request: https://github.com/dagster-io/dagster/pull/10788