[dagit] Clean up leftover escaping in plaintext from markdown (#10869)
### Summary & Motivation

Resolves #10331.

The remark plugin that converts markdown to plaintext leaves some escpaing backslashes behind. We don't need them for the plain string, so remove them.

### How I Tested These Changes

Modify an asset description to have `_ &asdf` as part of its text, load asset graph. Verify that the undescore and ampersand are not escaped in the rendered string.

Wrote a test for this as well.