dagster-graphql: external: gen_location_state_changes: state subscriber  bugfix
During the migration to Graphene 3, the `gen_location_state_changes`
function introduced a `_enqueue` function which was added to the
context's state subscribers. Mismatching the source of the subscribers
dictionary, a function was added instead of a `LocationStateSubscriber`
with the function as callback.
Detected the issue on our Dagster deployment as `AttributeError:
'function' object has no attribute 'handle_event'` on
`dagster/_core/workspace/context.py:492`
(`subscriber.handle_event(event)`).