[dagit] Use code location name instead of ID for toast (#10933)
### Summary & Motivation

The `id` field on code location entry status is prepended with `location_status`, which appears in the toast in Dagit when a code location is added.

Instead, use `name`, which is the appropriate display name of the code location. 

### How I Tested These Changes

Load dagit-cloud, add a code location. Verify that the toast has the `name` string.