[dagit] Use regular name@location for repo tag on runs (#10950)
### Summary & Motivation

In cases where we are using the RepoAddress values to query for a run tag, we should continue using `repo@location`. This is not expected to be URI-encoded.

This should fix an issue where launching a run does not include the dunder repo in the created tag.

### How I Tested These Changes

Run dagit with `-f` for a file that will use a dunder repo name. Launch a run. Verify that I see the (failed) run in the "Runs" tab on that job within Dagit, indicating that it can be queried successfully by tag.