[convert-development-to-production-7] Convert guide to use Definitions
Summary & Motivation:

This changes the development to production guide to correspond to the new Definitions API
and its improvments. Overall I think this feels pretty good. Sometimes there is a bit of awkwardness
where it says "in the Definitions object" rather than "in the repository". Also we may want to move
Definitions objects to a `defs.py` file instead of a `__init__.py` for clarity.

This depends on a diff stack that changes the underlying code.

Test Plan: Read

Reviewers: erinkcochran87

Reviewed By: erinkcochran87

Pull Request: https://github.com/dagster-io/dagster/pull/10744