[code-location-selector-stack Add CodeLocationSelector; Have run_status_sensor accept it
Summary & Motivation: Adds the ability to target a code location in the run_status_sensor rather than
a repository.

This is meant to be used in concert with the `Definitions` API that no longer exposes
the concept of a repository to a user. The new language is that the `Definitions` object determines the set
of definitions within a code location.

This diff hoists the computation of other_repos, other_repo_jobs, and current_repo_jobs out of the closure
so we get code coverage from the local unit test

Test Plan:
BK. Added additional testing in https://github.com/dagster-io/dagster/pull/10970

Reviewers: jamiedemaria

Reviewed By: jamiedemaria

Pull Request: https://github.com/dagster-io/dagster/pull/10783