Check scheduler ticks right after each minute boundary instead of once every 5 seconds (#10886)
Timing tweaks to try to reduce the time that the scheduler is sleeping before a tick

Summary:
A user reported that they were seeing between 1 and 6 seconds of latency when running the scheduler even when running a single simple schedule - this probably stems from our 5 second sleep interval. We have a simple tweak that we can do to try to maximize the chance that a tick will happen right after a 5-second boundary window.

The other change here is to include work that happens outside of the iterator in the timer (e.g. writing the heartbeat), which I did for sensors as well.

Test Plan:
Existing scheduler/sensor suite
Locally, run scheduler with a time.sleep(1) inside the heartbeat write function - see that ticks still happen every 5 seconds instead of every 6 seconds