Allow setting base_dir as a StringSource in sqlite storage
Summary:
We say you can do this in the docs but it doesn't work - this makes it work.

There's a gotcha though, we can't set the legacy storage fields if we do this since they derive their values from the base_dir. I think it's OK for back-compat to not work in this scenario (it seems worse than parsing the env var and putting its value in plaintext) but open to other possibilities.