Replace usages of `nslookup` with `nc` for user deployments
The effect of the two commands is basically the same, confirming the user deployments are functioning.  `nc` is slightly more robust as `nslookup` will fail if it fails on any of it's queries it takes, and apparently there are some cases where kubedns will be quirky and makes it spaz out.