Add a graphql field and API for location-scoped permissions
Summary:
Adds a permissions field to GrapheneRepoLocation that can be used to scope permissions to particular code locations. The idea is that the frontend will, for these permissions, instead of sourcing them for a single static map, source them from a map for each code location.

A thing we will need to figure out is how to handle permissions checks during edge cases where the underlying location is no longer in the workspace (for example, terminating a run after its code location has been removed from the workspace). One possibility to handle that case would be to add canXXX fields on the relevant graphql object that is then mutated. Another is to fall back to the global permissions map to check if they are permissioned to do it for *all* locations (which is what the server currently does).

Added an example handler for launching and re-exeucting a run to demonstrate what the serverside checks would look like.