Coerce IOManager objects into IOManagerDefinitions in Definitions
Summary & Motivation: Right now if you pass in a raw I/O manager object into the resources argument of Definitions it counterintutively is wrapped in a ResourceDefinition rather than an IOManagerDefinition. This change detects if an incoming object is an instance of IOManager and coerces it.

Test Plan: BK

Reviewers: yuhan, sryza

Reviewed By: sryza

Pull Request: https://github.com/dagster-io/dagster/pull/11107