Retry registering task definitions
When multiple runs try to register a task definition simultaneously, we
can hit an ECS API error:

```
botocore.errorfactory.ClientException: An error occurred (ClientException)
when calling the RegisterTaskDefinition operation: Too many concurrent
attempts to create a new revision of the specified family.
```

Our task definition reuse strategy creates a race condition where:

- multiple simultaneous runs can't find a task defintion
- run A successfully registers the task definition, run B hits the error

Now run B will retry and see that it can reuse the task definition that
run A created.

Testing this is proving to be tricky - I initially added a lock to our
stub and ran man concurrent threads against it. But getting the botocore
stubber interactions to work well with multiple threads was a challenge.
I'm putting this up for review now and I'll follow up with a better test
- likely one that makes heavy use of mock.patch but I'm open to
  suggestions.