[dagit] Optimizations to backfill UI for large partition key sets (#11201)
### Summary & Motivation

This PR addresses performance issues reported in
https://github.com/dagster-io/dagster/issues/11157

With a large set of partition keys (35000 in this example), the Dagit
backfill modal is unacceptably slow to load, and then also renders very
slowly as you interact with it. These tweaks make things more
reasonable: https://www.loom.com/share/c446d0f2123a45059b5182d533aca057

1) When you open the modal, we fetch the names of the partition
dimensions via `dimensionTypes[].name` rather than
`partitionKeysByDimension.name` which is much faster. I also extended
this backend API to work in the single dimension case so that it matches
the number of dimensions returned by partitionKeysByDimension (cc
@clairelin135 on this bit!)

2) Add a few optimizations for the "select all" case - when we see the
selected partition count and the all partition count are the same, we
can show `[A...Z]` fast.

3) Add a few useMemo's that were missing

4) Stop unnecessarily sorting the selected keys. We actually don't rely
on this anywhere as far as I can tell and it's very expensive.

5) For 35k items, it turns out it's quite a bit faster to convert the
Array to a Set and then use JS's `Set.has()`. Ideally we would convert
the selection state itself from an Array to a Set, but it would make
this PR much larger and the Set constructor is very very fast.

6) Update PartitionStatus.tsx to use a helper for building "spans" that
creates fewer intermediate data structures.

### How I Tested These Changes

Tested the backfill modal with existing assets in various states to make
sure the spans, etc. still behave the same way. Tested with an asset
with 35k partitions to see speed improvements.

Co-authored-by: bengotow <bgotow@elementl.com>