docs: update license to include year and company (#11231)
The year and company name were left out from when this was copied from a
template. This update just sets them so that the project gets proper
attribution in the license.

### Summary & Motivation
Was checking out the project for the first time and happened to notice
this in your LICENSE file.

### How I Tested These Changes
This change does not impact any code directly, but simply gives proper
author attribution in the LICENSE (IANAL).