Limit concurrent task definition revisions in stub
We already use a pretty complicated StubbedEcs for unit testing our ECS
code. It's not perfect but it lets us make a lot of useful assertions
without reaching out to the real API.

This complicates the stub further by making it callable from multiple
threads concurrently. This allows us to test previously untestable API
interactions like the limit on concurrent calls to revise the same task
definition family.

Two new tests were added: one to assert that the stub itself is working
as intended and another to assert that the logic introduced to fix our
race condition for registering tasks works.