[dagit] Utility for timezone-aware date/time formatting (#11285)
### Summary & Motivation

Create a utility hook for generating formatted date/time strings with
awareness of the user's timezone.

The hook returns a memoized function that accepts a date and an
`Intl.DateTimeOptions` object, and produces a formatted string with the
user's timezone applied to it.

I wrote a test and updated `RunTimeline` to make use of it. Most of the
use cases look like they'll be in Cloud.

### How I Tested These Changes

Jest, load Dagit and view run timeline. Verify that the date and time
labels look correct.