Pass duckdb_path to __init__ rather than relying on context (#11300)
### Summary & Motivation

In general we should not be relying on the context for values that are known, immutable, and typed at init time. Passing these directly to the initializer is more obvious and eases refactoring later down the line. More of the code ends up being not dependent on the context object and is more reusable.

### How I Tested These Changes

BK