backfill perf: swap backfill requested for num cancelable (#11304)
### Summary & Motivation
Context: this is the last step to bring backfill page loads to a
reasonable spot, after https://github.com/dagster-io/dagster/pull/11205
and https://github.com/dagster-io/dagster/pull/11246

Looking at a slow backfill page load on speedscope, it looks like we're
still doing slow multi-run fetches just to calculate the count of
"cancelable" runs, where "cancelable" is the number of partitions
requested in the backfill that never had corresponding runs launched.
This value was only used to cancel a backfill via the action menu.

Previously, we were comparing the number of runs with the backfill tag
on it to the total count of partitions requested.

In this PR, instead, we directly look at the checkpoint value on the
backfill object (e.g. how the backfill daemon checkpoints its own
progress in terms of requesting runs) and the backfill status to get a
`numCancelable` field. This has the advantage of never having to query
the runs table.

### How I Tested These Changes
Loaded slow backfill page in shadow dagit, saw the
`InstanceBackfillQuery` load time go from >15s down to 3s.

Launched a backfill, saw that I could successfully cancel.